bank_etl_v3/requirements.txt
pandas>=2.0.0
openpyxl>=3.1.2
python-dateutil>=2.9.0
PyYAML>=6.0.2
rapidfuzz>=3.6.1
xlsxwriter>=3.2.0
